requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5.2
boto3==1.13.9
botocore==1.16.9
colorama==0.4.3
Django==3.0.5
django-cors-headers==3.4.0
django-s3direct==1.1.5
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
dnspython==1.16.0
docutils==0.15.2
filelock==3.0.12
idna==2.9
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.5
py3-validate-email==0.2.9
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.15
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
Unidecode==1.1.1
urllib3==1.25.9
wrapt==1.11.2
